refactor(components): make card font module worker-safe
- Move Node TTF loader (monoFontUrls, loadCardFonts) out of fonts.ts into
  new fontsNode.ts so fonts.ts (imported by the card tree) has no node:*
  reference and bundles cleanly for browser + Cloudflare Worker
- Type SatoriFont.data as ArrayBuffer | Uint8Array (drop the Node Buffer
  global) so the module needs no @types/node
- Folder barrel re-exports fontsNode, keeping the public surface unchanged
